PIXHAWK: Enhanced FW loading error handling
This makes error handling during FW flashing more robust, preventing most cases of locked out bricking. Adds a 120 second timeout to the px_uploader subprocess, which will allow the IMX to continue booting up, providing WiFi access to the file system to correct the problem.  If firmware loading times out or returns an error code, the offending FW file will be deleted from the /firmware/ directory. This prevents the IMX from continuously trying to load it on every reboot. If the pixhawk is stuck in bootloader, or otherwise stuck without operational FW, this will restore WiFi SSH/SFTP access to the IMX to recover the mess.  Before, you were stuff with an inaccessible brick that needed to be factory reset and disassembled.