PX_UPLOADER: Enhanced error handling and reporting
This patch allows px_uploader to properly return 0 if flashing succeeded or return 1 if flashing did not succeed.  It will also print any error instead of only one type of error.  For systems with px_uploader being run automatically (such as on a Solo), this will allow the calling scripts to know if loading succeeded or not.  Previously, px_uploader exited with 0 regardless or success or failure.